调整PWM+DMA配置，实现灯板彩虹流水灯的效果，编写注意事项

+++ Hardware/WS2812.h
@@ -7,8 +7,8 @@
 
 #include "main.h"
 
-void WS2812_Code(uint8_t Red, uint8_t Green, uint8_t Blue, uint16_t *color_byte);
-extern uint16_t LED_Color_Byte[];
-extern uint16_t LED_Zero_Byte[];
+void WS2812_Code(uint8_t Red, uint8_t Green, uint8_t Blue, uint32_t *color_byte);
+extern uint32_t LED_Color_Byte[];
+extern uint32_t LED_Zero_Byte[];
 
 #endif //WS2812_WS2812_H
